Push shared logic into base class
