+++ src/dodal/devices/beamlines/i15_1/cryostream.py
@@ -1,29 +1,10 @@
-from daq_config_server import ConfigClient
 from daq_config_server.models.i15_1.xpdf_parameters import (
     TemperatureControllerParams,
-    TemperatureControllersConfig,
 )
 
-from dodal.devices.beamlines.i15_1.temperature_controller import TemperatureController
+from dodal.devices.beamlines.i15_1.temperature_controller import SafeOrBeamPositioner
 
 
-class Cryostream(TemperatureController):
-    def __init__(self, prefix: str, config_client: ConfigClient, xpdf_parameters_path):
-        self.config_client = config_client
-        self.xpdf_parameters_path = xpdf_parameters_path
-        super().__init__(prefix)
-
+class Cryostream(SafeOrBeamPositioner):
     def get_config(self) -> TemperatureControllerParams:
-        return self.config_client.get_file_contents(
-            self.xpdf_parameters_path,
-            desired_return_type=TemperatureControllersConfig,
-            force_parser=TemperatureControllersConfig.from_xpdf_parameters,
-        ).cryostream
-
-    @property
-    def _safe_position(self) -> float:
-        return self.get_config().safe_position
-
-    @property
-    def _beam_position(self) -> float:
-        return self.get_config().beam_position
+        return self.get_full_config().cryostream
